Method 4: Concurrent processing with file chunks
Instead of the line-by-line concurrent processing of method 3, which clearly spawned too many processes to be performant in any way, this method splits the source file in equal chunks. Then, for every chunk a goroutine is spawned to scan the contents and parse the station information. This method shows a significant performance improvement.